fix: Move extractMLTrainingData to separate utility file
Server Actions files ('use server') require all exported functions to be async.
Move helper function to src/lib/ml/dataExtractor.ts to resolve build error.

- Create src/lib/ml/dataExtractor.ts
- Move MLTrainingData type and extractMLTrainingData function
- Update imports in mlDataCollectionActions.ts and gameLogicActions.ts
